+++ lib/beacon/web/beacon_assigns.ex
@@ -33,8 +33,8 @@ defmodule Beacon.Web.BeaconAssigns do
         }
 
   defstruct site: nil,
-            path_params: %{},
-            query_params: %{},
+            path_params: nil,
+            query_params: nil,
             page: %{path: nil, title: nil},
             private: %{
               page_module: nil,
@@ -53,31 +53,18 @@ defmodule Beacon.Web.BeaconAssigns do
   end
 
   @doc false
-  def new(site, %Beacon.Content.Page{} = page, variant_roll) do
-    components_module = Beacon.Loader.Components.module_name(site)
-    page_module = Beacon.Loader.Page.module_name(site, page.id)
+  def new(%Beacon.Content.Page{} = page, metadata \\ []) do
+    path_info = Keyword.get(metadata, :path_info, [])
+    query_params = Keyword.get(metadata, :query_params, %{})
+    variant_roll = Keyword.get(metadata, :variant_roll, nil)
 
-    %__MODULE__{
-      site: site,
-      private: %{
-        components_module: components_module,
-        page_module: page_module,
-        variant_roll: variant_roll
-      }
-    }
-  end
-
-  @doc false
-  def new(site, %Beacon.Content.Page{} = page, live_data, path_info, query_params, source, variant_roll \\ nil)
-      when is_atom(site) and is_map(live_data) and is_list(path_info) and is_map(query_params) and source in [:beacon, :admin] do
-    %{site: ^site} = page
-    page_module = Beacon.Loader.Page.module_name(site, page.id)
-    live_data = Beacon.Web.DataSource.live_data(site, path_info, Map.drop(query_params, ["path"]))
+    page_module = Beacon.Loader.Page.module_name(page.site, page.id)
+    live_data = Beacon.Web.DataSource.live_data(page.site, path_info, Map.drop(query_params, ["path"]))
     path_params = Beacon.Router.path_params(page.path, path_info)
-    page_title = Beacon.Web.DataSource.page_title(site, page.id, live_data, source)
-    components_module = Beacon.Loader.Components.module_name(site)
-    info_handlers_module = Beacon.Loader.InfoHandlers.module_name(site)
-    event_handlers_module = Beacon.Loader.EventHandlers.module_name(site)
+    page_title = Beacon.Web.DataSource.page_title(page, live_data)
+    components_module = Beacon.Loader.Components.module_name(page.site)
+    info_handlers_module = Beacon.Loader.InfoHandlers.module_name(page.site)
+    event_handlers_module = Beacon.Loader.EventHandlers.module_name(page.site)
 
     %__MODULE__{
       site: page.site,
@@ -94,24 +81,5 @@ defmodule Beacon.Web.BeaconAssigns do
         variant_roll: variant_roll
       }
     }
-  end
-
-  @doc false
-  def update(_socket_or_assigns, _key, _value)
-
-  def update(%{assigns: %{beacon: _beacon}} = socket, key, value) do
-    do_update_socket_or_assigns(socket, key, value)
-  end
-
-  def update(%{beacon: _beacon} = assigns, key, value) do
-    do_update_socket_or_assigns(assigns, key, value)
-  end
-
-  def update(_socket_or_assigns, _key, _value), do: raise("expected :beacon assign in socket but none found")
-
-  defp do_update_socket_or_assigns(socket_or_assigns, key, value) do
-    Phoenix.Component.update(socket_or_assigns, :beacon, fn beacon ->
-      Map.put(beacon, key, value)
-    end)
   end
 end