+++ lib/beacon/web/data_source.ex
@@ -3,12 +3,14 @@ defmodule Beacon.Web.DataSource do
 
   require Logger
 
-  def live_data(site, path_info, params \\ %{}) when is_atom(site) and is_list(path_info) and is_map(params) do
-    Beacon.apply_mfa(site, Beacon.Loader.fetch_live_data_module(site), :live_data, [path_info, params])
+  def live_data(site, path_info, query_params) when is_atom(site) and is_list(path_info) and is_map(query_params) do
+    Beacon.apply_mfa(site, Beacon.Loader.fetch_live_data_module(site), :live_data, [path_info, query_params])
   end
 
-  def page_title(site, page_id, live_data, source) do
-    %{path: path, title: title} = page_assigns = page_assigns(site, page_id, source)
+  def live_data(_site, _path_info, _query_params), do: %{}
+
+  def page_title(%Beacon.Content.Page{} = page, live_data) do
+    %{path: path, title: title} = page_assigns = page_assigns(page)
 
     with {:ok, page_title} <- Beacon.Content.render_snippet(title, %{page: page_assigns, live_data: live_data}) do
       page_title
@@ -19,7 +21,7 @@ defmodule Beacon.Web.DataSource do
 
         will return the original unmodified page title
 
-        site: #{site}
+        site: #{page.site}
         title: #{title}
         page path: #{path}
 
@@ -36,13 +38,14 @@ defmodule Beacon.Web.DataSource do
   # TODO: revisit this logic to evaluate meta_tags for unpublished pages
   def meta_tags(assigns) do
     %{beacon: %{site: site, private: %{page_module: page_module, live_data_keys: live_data_keys}}} = assigns
-    %{site: ^site, id: page_id} = Beacon.apply_mfa(site, page_module, :page_assigns, [[:site, :id]])
+    %{site: ^site} = page_assigns = Beacon.apply_mfa(site, page_module, :page_assigns, [])
+
     live_data = Map.take(assigns, live_data_keys)
 
     assigns
     |> Beacon.Web.Layouts.meta_tags()
     |> List.wrap()
-    |> Enum.map(&interpolate_meta_tag(&1, %{page: page_assigns(site, page_id, :beacon), live_data: live_data}))
+    |> Enum.map(&interpolate_meta_tag(&1, %{page: page_assigns, live_data: live_data}))
   end
 
   defp interpolate_meta_tag(meta_tag, values) when is_map(meta_tag) do
@@ -57,15 +60,21 @@ defmodule Beacon.Web.DataSource do
     end
   end
 
-  # only published pages will have the title evaluated for beacon
-  defp page_assigns(site, page_id, :beacon) do
-    Beacon.apply_mfa(site, Beacon.Loader.fetch_page_module(site, page_id), :page_assigns, [])
+  # return the page assigns from unpublished page assigns,
+  # either saved in the database or page in-memory (for new pages)
+  # this fallback is here mostly to support  beacon_live_admin visual editor,
+  # which makes use of the `@beacon` assign when creating or editing pages
+  defp page_assigns(%Beacon.Content.Page{id: nil} = page) do
+    unpublished_page_assigns(page)
   end
 
-  # beacon_live_admin needs the title for unpublished pages also
-  defp page_assigns(site, page_id, :admin) do
-    page = Beacon.Content.get_page(site, page_id)
+  defp page_assigns(%Beacon.Content.Page{} = page) do
+    Beacon.apply_mfa(page.site, Beacon.Loader.fetch_page_module(page.site, page.id), :page_assigns, [])
+  rescue
+    _ -> unpublished_page_assigns(page)
+  end
 
+  defp unpublished_page_assigns(page) do
     %{
       id: page.id,
       site: page.site,