+++ CHANGELOG.md
@@ -6,12 +6,23 @@
   - Beacon will now automatically generate a `sitemap.xml` for each `beacon_site` defined in the Router
   - Add macro `beacon_sitemap_index` for use in the Router to serve a sitemap index
 
+## 0.3.3 (2024-12-13)
+
+### Fixes
+  - Support LiveView v1.0.1
+  - Fix variant roll changing when fetching assets
+
+## 0.3.2 (2024-12-11)
+
+### Fixes
+  - Make the logic to find reachable sites less strict
+  - Prevent components module reloading in manual mode
+
 ### Doc
   - Add missing instructions on Upgrade Guide v0.3.0 where to place the Beacon tuple
 
 ### Chore
   - Exclude Dialyzer files from package
-
 
 ## 0.3.1 (2024-12-10)